+++ modelo/Gerente.java
@@ -8,7 +8,7 @@ public class Gerente extends Funcionario {
 
 	
 	public Gerente(int id, long cpf, String nome, String sobrenome, String nomeDeUsuario, String pais, String estado,
-			String cidade, String rua, String bairro, int cep, int numeroDaResidencia, byte ddi, byte ddd, int telefone,
+			String cidade, String rua, String bairro, int cep, short numeroDaResidencia, byte ddi, byte ddd, int telefone,
 			boolean ativo, boolean superGerente) {
 		super(id, cpf, nome, sobrenome, nomeDeUsuario, pais, estado, cidade, rua, bairro, cep, numeroDaResidencia, ddi,
 				ddd, telefone, ativo);
@@ -17,7 +17,7 @@ public class Gerente extends Funcionario {
 
 	
 	public Gerente(long cpf, String nome, String sobrenome, String nomeDeUsuario, String pais, String estado,
-			String cidade, String rua, String bairro, int cep, int numeroDaResidencia, byte ddi, byte ddd, int telefone,
+			String cidade, String rua, String bairro, int cep, short numeroDaResidencia, byte ddi, byte ddd, int telefone,
 			boolean ativo, boolean superGerente) {
 		super(cpf, nome, sobrenome, nomeDeUsuario, pais, estado, cidade, rua, bairro, cep, numeroDaResidencia, ddi,
 				ddd, telefone, ativo);
